beacon state
