+++ src/reducers/app/actions.ts
@@ -14,7 +14,6 @@ import {
     SetPendingOperationsAction,
 } from './types';
 import {Operation} from '../types';
-import {BeaconErrorMessage} from '../../beacon/types';
 
 export const SET_KEYS = 'SET_KEYS';
 export const SET_BALANCE = 'SET_BALANCE';
@@ -28,36 +27,6 @@ export const SET_DELEGATE_ADDRESS = 'SET_DELEGATE_ADDRESS';
 export const SET_DELEGATION = 'SET_DELEGATION';
 export const SET_EXPECTEDDELEGATEDATE = 'SET_EXPECTEDDELEGATEDATE';
 export const SET_PENDING_OPERATIONS = 'SET_PENDING_OPERATIONS';
-export const SET_BEACON_MESSAGE = 'SET_BEACON_MESSAGE';
-export const SET_BEACON_PERMISSIONS = 'SET_BEACON_PERMISSIONS';
-export const SET_BEACON_METADATA = 'SET_BEACON_METADATA';
-export const SET_BEACON_PERMISSIONS_LOADING = 'SET_BEACON_PERMISSIONS_LOADING';
-export const SET_BEACON_ERROR_MESSAGE = 'SET_BEACON_ERROR_MESSAGE';
-
-export const setBeaconMessage = (beaconMessage: any = {}) => ({
-    type: SET_BEACON_MESSAGE,
-    beaconMessage
-});
-
-export const setBeaconPermissions = (beaconPermissions: any = []) => ({
-    type: SET_BEACON_PERMISSIONS,
-    beaconPermissions,
-});
-
-export const setBeaconMetadata = (beaconMetadata: any = []) => ({
-    type: SET_BEACON_METADATA,
-    beaconMetadata,
-});
-
-export const setBeaconPermissionsLoading = (beaconPermissionLoading: boolean = false) => ({
-    type: SET_BEACON_PERMISSIONS_LOADING,
-    beaconPermissionLoading,
-});
-
-export const setBeaconErrorMessage = (beaconErrorMessage: BeaconErrorMessage | null = null) => ({
-    type: SET_BEACON_ERROR_MESSAGE,
-    beaconErrorMessage,
-});
 
 export const setKeysAction = (keys: KeyStore): SetKeysAction => ({
     type: SET_KEYS,
@@ -89,7 +58,9 @@ export const setSendAmount = (amount: number): SetSendAmountAction => ({
     amount,
 });
 
-export const setTransactions = (transactions: Operation[]): SetTransactionsAction => ({
+export const setTransactions = (
+    transactions: Operation[],
+): SetTransactionsAction => ({
     type: SET_TRANSACTIONS,
     transactions,
 });
@@ -111,13 +82,18 @@ export const setDelegation = (delegation: string): SetDelegationAction => ({
     delegation,
 });
 
-export const setDelegateExpectedDate = (date: Date): SetDelegationExpectedDate => ({
+export const setDelegateExpectedDate = (
+    date: Date,
+): SetDelegationExpectedDate => ({
     type: SET_EXPECTEDDELEGATEDATE,
-    date
+    date,
 });
 
-export const setPendingOperations = (transactions: any[], delegations: any[]): SetPendingOperationsAction => ({
+export const setPendingOperations = (
+    transactions: any[],
+    delegations: any[],
+): SetPendingOperationsAction => ({
     type: SET_PENDING_OPERATIONS,
     transactions,
-    delegations
-})
+    delegations,
+});
