+++ src/reducers/app/reducers.ts
@@ -13,11 +13,6 @@ import {
     SET_DELEGATION,
     SET_EXPECTEDDELEGATEDATE,
     SET_PENDING_OPERATIONS,
-    SET_BEACON_MESSAGE,
-    SET_BEACON_PERMISSIONS,
-    SET_BEACON_METADATA,
-    SET_BEACON_PERMISSIONS_LOADING,
-    SET_BEACON_ERROR_MESSAGE,
 } from './actions';
 
 const initialState = {
@@ -38,23 +33,13 @@ const initialState = {
     expectedPaymentDate: null,
     pendingTransactions: [],
     pendingDelegations: [],
-    beaconMessage: {},
-    beaconPermissions: [],
-    beaconMetadata: [],
-    beaconPermissionLoading: false,
-    beaconErrorMessage: null,
 };
 
 const app = (state = initialState, action: AppActions) => {
     switch (action.type) {
         case SET_KEYS:
-            const {
-                publicKey,
-                secretKey,
-                publicKeyHash,
-                storeType,
-                seed,
-            } = action.keys;
+            const {publicKey, secretKey, publicKeyHash, storeType, seed} =
+                action.keys;
             return {
                 ...state,
                 publicKey,
@@ -84,17 +69,11 @@ const app = (state = initialState, action: AppActions) => {
         case SET_EXPECTEDDELEGATEDATE:
             return {...state, expectedPaymentDate: action.date};
         case SET_PENDING_OPERATIONS:
-            return {...state, pendingTransactions: action.transactions, pendingDelegations: action.delegations};
-        case SET_BEACON_MESSAGE:
-            return {...state, beaconMessage: action.beaconMessage};
-        case SET_BEACON_PERMISSIONS:
-            return {...state, beaconPermissions: action.beaconPermissions};
-        case SET_BEACON_METADATA:
-            return {...state, beaconMetadata: action.beaconMetadata};
-        case SET_BEACON_PERMISSIONS_LOADING:
-            return {...state, beaconPermissionLoading: action.beaconPermissionLoading};
-        case SET_BEACON_ERROR_MESSAGE:
-            return {...state, beaconErrorMessage: action.beaconErrorMessage};
+            return {
+                ...state,
+                pendingTransactions: action.transactions,
+                pendingDelegations: action.delegations,
+            };
         default:
             return state;
     }
