+++ src/reducers/app/types.ts
@@ -12,15 +12,8 @@ import {
     SET_DELEGATION,
     SET_EXPECTEDDELEGATEDATE,
     SET_PENDING_OPERATIONS,
-    SET_BEACON_MESSAGE,
-    SET_BEACON_PERMISSIONS,
-    SET_BEACON_METADATA,
-    SET_BEACON_PERMISSIONS_LOADING,
-    SET_BEACON_ERROR_MESSAGE,
 } from './actions';
 import {Operation} from '../types';
-import {BeaconErrorMessage} from '../../beacon/types';
-
 export interface State {
     publicKey: string; // TODO: store a KeyStore object
     secretKey: string;
@@ -44,36 +37,7 @@ export interface State {
     expectedPaymentDate: Date;
     pendingTransactions: [];
     pendingDelegations: [];
-    beaconMessage: any;
-    beaconPermissions: any;
-    beaconMetadata: any;
-    beaconPermissionLoading: boolean;
-    beaconErrorMessage: BeaconErrorMessage | null;
 }
-
-export interface BeaconErrorMessageAction {
-    type: typeof SET_BEACON_ERROR_MESSAGE;
-    beaconErrorMessage: BeaconErrorMessage | null;
-}
-export interface BeaconPermissionLoadingAction {
-    type: typeof SET_BEACON_PERMISSIONS_LOADING;
-    beaconPermissionLoading: boolean;
-}
-export interface BeaconMessageAction {
-    type: typeof SET_BEACON_MESSAGE;
-    beaconMessage: any;
-}
-
-export interface BeaconPermissionsAction {
-    type: typeof SET_BEACON_PERMISSIONS;
-    beaconPermissions: any;
-}
-
-export interface BeaconMetadataAction {
-    type: typeof SET_BEACON_METADATA;
-    beaconMetadata: any
-}
-
 export interface SetKeysAction {
     type: typeof SET_KEYS;
     keys: KeyStore;
@@ -131,8 +95,8 @@ export interface SetDelegationExpectedDate {
 
 export interface SetPendingOperationsAction {
     type: typeof SET_PENDING_OPERATIONS;
-    transactions: any[],
-    delegations: any[]
+    transactions: any[];
+    delegations: any[];
 }
 
 export type AppActions =
@@ -147,9 +111,4 @@ export type AppActions =
     | SetDelegateAddressAction
     | SetDelegationAction
     | SetDelegationExpectedDate
-    | SetPendingOperationsAction
-    | BeaconMessageAction
-    | BeaconPermissionsAction
-    | BeaconMetadataAction
-    | BeaconPermissionLoadingAction
-    | BeaconErrorMessageAction;
+    | SetPendingOperationsAction;
